Refactor DataCite functions
